25/08/10 [feat] 데미지 UI

+++ Castawild/Assets/Scripts/CHH_Scripts/Object/AttackObject.cs
@@ -2,6 +2,6 @@ using Fusion;
 
 public class AttackObject : NetworkBehaviour
 {
-    public float att;
+    public int att;
     public bool canAttack;
 }
